% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/launch_app.R
\name{launch_app}
\alias{launch_app}
\title{Launch Shiny App}
\usage{
launch_app()
}
\value{
Runs the Shiny app in the web browser.
}
\description{
Runs the Shiny app located in the "shiny" directory of the package "TopicsModelingTools".
}
\examples{
launch_app()

}
